backend/requirement.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==1.4.48
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.1.0